% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanGutMicrobiomeData.R
\docType{data}
\name{ad}
\alias{ad}
\title{Data of Zhuang et al's 2018 study examining fecal samples representing the gut microbiome of patients with a clinical diagnosis of Alzheimer's disease and age-sex matched controls in normalized proportions. Data were reprocessed in dada2 using Silva database 138.2.}
\format{
A phyloseq object:
\describe{
\item{otu_table}{Relative abundance of ASVs per each sample. 16359 ASVs are represented across 86 samples}
\item{sample_data}{Metadata table connecting metadata to each sample. Contains 86 samples by 27 sample variables}
\item{tax_table}{A table listing the taxonomic assignment of each ASV. Contains 16359 taxa across 7 taxonomic ranks}
}
}
\source{
\url{https://doi.org/10.3233/JAD-180176}
}
\usage{
ad
}
\description{
Relative abundance data for fecal samples representing the gut microbiome of subjects with a clinical diagnosis of Alzheimer's disease and age-sex matched controls.
}
\keyword{datasets}
