% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanGutMicrobiomeData.R
\docType{data}
\name{fefifo}
\alias{fefifo}
\title{Data of Sonnenburg Lab's FeFiFo study in normalized proportions}
\format{
A phyloseq object:
\describe{
\item{otu_table}{Relative abundance of ASVs per each sample. 2134 ASVs are represented across 311 samples}
\item{sample_data}{Metadata table connecting metadata to each sample. Contains 311 samples by 27 sample variables}
\item{tax_table}{A table listing the taxonomic assignment of each ASV. Contains 2134 taxa across 7 taxonomic ranks}
}
}
\source{
\url{https://github.com/SonnenburgLab/fiber-fermented-study}
}
\usage{
fefifo
}
\description{
Relative abundance data for fecal samples representing the gut microbiome of subjects on fiber-rich and fermented food diets
}
\keyword{datasets}
