% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanGutMicrobiomeData-package.R
\docType{package}
\name{HumanGutMicrobiomeData-package}
\alias{HumanGutMicrobiomeData}
\alias{HumanGutMicrobiomeData-package}
\title{HumanGutMicrobiomeData: Data Package for Various Human Gut 16S studies}
\description{
Contains the RDS files and code used to generate them for various human gut 16S rRNA amplicon data meant for use in the C-MOOR Human Gut 16srRNA Microbiome miniCURE.
}
\author{
\strong{Maintainer}: Sayumi York \email{syork@ndm.edu}

}
\keyword{internal}
