% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanGutMicrobiomeData.R
\docType{data}
\name{oc_agefirstOCuse_counts}
\alias{oc_agefirstOCuse_counts}
\title{Data of Terrazas et al.'s 2025 study on the effect of oral contraceptives on the gut microbiome, in raw counts. Only the samples with data for agefirstOCuse are included}
\format{
A phyloseq object:
\describe{
\item{otu_table}{Counts of ASVs per each sample. 5457 ASVs are represented across 89 samples}
\item{sample_data}{Metadata table connecting metadata to each sample. Contains 89 samples by 25 sample variables}
\item{tax_table}{A table listing the taxonomic assignment of each ASV. Contains 5457 taxa across 7 taxonomic ranks}
}
}
\source{
\url{https://github.com/fernanda-t/microbiome-oc}
}
\usage{
oc_agefirstOCuse_counts
}
\description{
Count data for fecal samples representing the gut microbiome of subjects on various dosages including none of oral contraceptives. Only the samples with data for agefirstOCuse are included
}
\keyword{datasets}
