% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanGutMicrobiomeData.R
\docType{data}
\name{oc_dosage}
\alias{oc_dosage}
\title{Data of Terrazas et al.'s 2025 study on the effect of oral contraceptives on the gut microbiome in normalized proportions. Only participants with dosage data are included.}
\format{
A phyloseq object:
\describe{
\item{otu_table}{Relative abundance of ASVs per each sample. 5457 ASVs are represented across 73 samples}
\item{sample_data}{Metadata table connecting metadata to each sample. Contains 73 samples by 25 sample variables}
\item{tax_table}{A table listing the taxonomic assignment of each ASV. Contains 5457 taxa across 7 taxonomic ranks}
}
}
\source{
\url{https://github.com/fernanda-t/microbiome-oc}
}
\usage{
oc_dosage
}
\description{
Relative abundance data for fecal samples representing the gut microbiome of subjects on various dosages including none of oral contraceptives. Only participants with dosage data are included.
}
\keyword{datasets}
