% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanGutMicrobiomeData.R
\docType{data}
\name{oc_no948N2_counts}
\alias{oc_no948N2_counts}
\title{Data of Terrazas et al.'s 2025 study on the effect of oral contraceptives on the gut microbiome in count data excluding sample 948N2}
\format{
A phyloseq object:
\describe{
\item{otu_table}{Counts of ASVs per each sample. 5457 ASVs are represented across 114 samples}
\item{sample_data}{Metadata table connecting metadata to each sample. Contains 114 samples by 25 sample variables}
\item{tax_table}{A table listing the taxonomic assignment of each ASV. Contains 5457 taxa across 7 taxonomic ranks}
}
}
\source{
\url{https://github.com/fernanda-t/microbiome-oc}
}
\usage{
oc_no948N2_counts
}
\description{
Count data for fecal samples representing the gut microbiome of subjects on various dosages including none of oral contraceptives, excluding sample 948N2
}
\keyword{datasets}
